refactor: move AppBar component from shared to app-development, since its only used there, and rewrite to fuction component
