+++ src/studio/src/designer/frontend/app-development/layout/appBarConfig.tsx
